Moved services code to seperate file
Basically anything with an api key was moved into the `service.py` file. This
will help us later on when we want to not store the api keys in the code itself, and
has the added benifit of decoupling some of the twilio and google api code. By
decoupling the twilio and google code, we can now easily switch out or turn off those
services which will come in handy for testing/developement purposes.