Set default host for apiclient to local
Clearing the database will only clear a locally running database so
clearing the production db is no more.

If apiclient is needed to be run on production, the best thing to
do would be to ssh into one of the production servers and run apiclient
locally from within the api docker container.