Added exception handling support
Also added a special `GatorException` that will be very usefull,
once refactoring makes it way to flask endpoints. Most of
logic will be moved outside of the flask endpoints so it is going
to be harder handle bad inputs/requests since the logic will be
spread out. It makes a lot of sense to handle this exceptional logic
with exceptions instead of conditionals