Implemented authentication and fb login
Logins require a facebook `fbuser_id` and `fbuser_token` and upon
successful login, a token is returned that can be sent to the api
to fullfill requests. Each endpoint has a list of permissions
that a client is required to have before it can access the resource,
which is verified by the `validate_permission` function.

closes #47
closes #48