Added local psudeo lambda and dynamo streams
The notification system does not work locally due to it's reliance
on aws lambda and dynamo streams. This commit imitates that
functionality by placing a reverse proxy (dynamoproxy.py) in
front of the local dynamodb that looks for dynamodb writes. When
a write is detected, it packages the write into a stream record
then sends it to the psuedo lambda server (lambdaserver.js). The
lambda server then executes the lambda code.

This is a very limited implementation. It currently does not support:
batch writes, the dynamo "UpdateExpression" field, multiple tables,
or multiple lambda functions.

Tests will be added later

track #76