"env.py start" pulls and runs the correct image
The behavior of "env.py start" has changed a little bit.
It will look in the Dockerrun.aws.json files to see what docker
image the api should be running. It then pulls this image from
Docker Hub, re-creates the containers, and starts them. This
means that the container's previous state is deleted upon
calling "env.py start".

Sometimes you won't want to start the api with the image specified
in the Dockerrun.aws.json file, instead you want to just start the
already create containers (regardless of whether they are based off
the image specified in the Dockerrun.aws.json). To do this, pass
in the `--local` flag and "env.py start" will *just* start the
containers.